Refactor & tidy patches against rdf-ldp
