+++ app/ladder.rb
@@ -1,15 +1,10 @@
 require 'rack/ldp'
 require 'sinatra/base'
 require 'rdf/mongo'
-# TODO: to remove
-require 'mongoid'
+require 'ladder/resource'
 
 # for debugging
 require 'pry'
-
-require_relative '../lib/ladder/rdf_source'
-require_relative '../lib/ladder/resource'
-# require_relative '../lib/ladder/non_rdf_source'
 
 module Ladder
   class LDP < Sinatra::Base
@@ -27,10 +22,6 @@ module Ladder
       # Use a class that implements the RDF::Repository interface
       set :uri, 'mongodb://localhost:27017/ladder'
       set :repository, RDF::Mongo::Repository.new(uri: uri)
-
-      # Configuration settings for Mongoid
-      # TODO: move to rdf_source.rb?
-      Mongoid.load_configuration({ clients: { default: { uri: uri } } })
     end
 
     get '/*' do
